requirements.txt
pandas~=2.2.3
numpy~=1.26.4
opencv-python~=4.10.0.84
matplotlib~=3.9.2
scikit-learn~=1.5.2
python-dotenv~=1.0.1
pytesseract~=0.3.13
pdf2image~=1.17.0
typing~=3.7.4.3
pip~=24.3.1
distro~=1.9.0
wheel~=0.45.1
protobuf~=5.29.3
cryptography~=43.0.1
pillow~=10.4.0
filelock~=3.16.1
pdfplumber~=0.11.4
scikit-image~=0.24.0
geopy~=2.4.1
Flask~=3.1.0
Flask-Session
Werkzeug~=3.1.3
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.4
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.3.0
networkx==3.4.2
numpy==1.26.4
openai==1.60.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tomli==2.0.1
tqdm==4.67.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.45.1
PyMuPDF==1.24.2
pywin32==306
PyPDF2~=3.0.1
mistralai>=0.1.6
